+++ CHANGELOG.md
@@ -4,11 +4,17 @@ All notable changes to this project will be documented in this file
 
 Log format based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [3.2.10] - 2021-11-19
+
+### Added:
+
+* `MyLab` console log formatter 
+
 ## [3.1.9] - 2021-09-29
 
 ### Changed
 
-* Short serialization of `System.Reflection.*` classes
+* Short serialization of `System.Reflection.*` classes	
 
 ## [3.1.8] - 2021-08-29